fix(voucher): enforce canonical CIP-68 metadata
