+++ cardano/gateway/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@cardano-ibc/planner": "file:../../packages/cardano-ibc-planner",
+    "@cardano-ibc/trace-registry": "file:../../packages/cardano-ibc-trace-registry",
     "@cardano-ibc/tx-builder": "file:../../packages/cardano-ibc-tx-builder",
     "@cardano-ogmios/client": "6.13.0",
     "@cardano-ogmios/schema": "6.13.0",
@@ -103,6 +104,8 @@
     ],
     "rootDir": "src",
     "moduleNameMapper": {
+      "^@cardano-ibc/trace-registry$": "<rootDir>/../../../packages/cardano-ibc-trace-registry/src/index.ts",
+      "^@cardano-ibc/trace-registry/(.*)$": "<rootDir>/../../../packages/cardano-ibc-trace-registry/src/$1",
       "^src/(.*)$": "<rootDir>/$1",
       "^~@/(.*)$": "<rootDir>/$1",
       "^@shared/(.*)$": "<rootDir>/shared/$1",
