+++ cardano/gateway/src/api/api.controller.modern.spec.ts
@@ -477,9 +477,9 @@ describe('ApiController (modern)', () => {
       voucher_policy_id: voucherPolicyId.toUpperCase(),
       ibc_denom_hash: 'c'.repeat(64),
       cip68_reference_asset_id: `${voucherPolicyId}${`000643b0${'b'.repeat(56)}`}`,
-      name: 'ATOM (IBC)',
+      name: 'uatom',
       description: 'IBC voucher for transfer/channel-7/uatom',
-      ticker: 'ATOM',
+      ticker: 'uatom',
     });
 
     const response = await controller.getCardanoAssetDenomTrace(`${voucherPolicyId}${voucherTokenName}`);
@@ -495,11 +495,11 @@ describe('ApiController (modern)', () => {
       cip68_reference_asset_id: `${voucherPolicyId}${`000643b0${'b'.repeat(56)}`}`,
       voucher_policy_id: voucherPolicyId.toUpperCase(),
       ibc_denom_hash: 'c'.repeat(64),
-      display_name: 'ATOM (IBC)',
-      display_symbol: 'ATOM',
+      display_name: 'uatom',
+      display_symbol: 'uatom',
       display_description: 'IBC voucher for transfer/channel-7/uatom',
       description: 'IBC voucher for transfer/channel-7/uatom',
-      ticker: 'ATOM',
+      ticker: 'uatom',
       decimals: null,
       url: null,
       logo: null,
@@ -552,7 +552,7 @@ describe('ApiController (modern)', () => {
         voucher_policy_id: 'f'.repeat(56),
         ibc_denom_hash: '1'.repeat(64),
         cip68_reference_asset_id: `${'f'.repeat(56)}${`000643b0${'e'.repeat(56)}`}`,
-        name: '1 (IBC)',
+        name: '1',
         description: 'IBC voucher for transfer/channel-3/gamm/pool/1',
         ticker: '1',
       },
@@ -572,7 +572,7 @@ describe('ApiController (modern)', () => {
         cip68_reference_asset_id: `${'f'.repeat(56)}${`000643b0${'e'.repeat(56)}`}`,
         voucher_policy_id: 'f'.repeat(56),
         ibc_denom_hash: '1'.repeat(64),
-        display_name: '1 (IBC)',
+        display_name: '1',
         display_symbol: '1',
         display_description: 'IBC voucher for transfer/channel-3/gamm/pool/1',
         description: 'IBC voucher for transfer/channel-3/gamm/pool/1',
